Major refactoring; Clean up; More features
Adding a player object, which holds the state of the player. Later it shall get methods to change the state of the player.
Adding an unfinished Logger class, which is in charge of formatting messages send to console.
Moving a lot of network classes into their own packages.
Storing Players instead of connections in the Core. Only when the Player has successfully joined the server. Sending example chunks from the core. Sending chat to other players.
Handshake stage creates and object with the client data and forwards it to the next stages.
Formatted serveral console messages.
Getting player amount in status message from the core.
Adding compression to the connection currently 1024 as threshold.
Removing unused methods from the cryption classes, adding a specific exception.
In PacketReader removing redundant method for reading a byte. Always read an unsigned byte. Sorting methods in the PacketBuilder.
Simple validation of the AuthServer response (check if empty, check is id and name and properties are set). Else disconnect.
Adding a flush and close method in the connection. And several catch clauses which are a bit more descriptive. No good shutdown management yet.
Adding more descriptors to the GameHandler.