Fix enableCompression; Adding SendHelper
Compression had been enabled before the compression packet had been sent. Causing it to be compressed and the client failing to read it.
The fix for that is to let the read thread wait for all packets to be sent until proceeding.
It may fail in some edge cases, to check checked again.
Changed some exception handling.