Adding blockdata/item resources (ID lookups)
The minecraft client will send itemID's and expects blockID's.
This commit adds the raw blockdata and item files from the original MC server.
Additional a manually created file has been added which states which item causes which block to be placed. This should be managed by sourcecode later on, especially since special blocks e.g. redstone are not placed correctly.
That file can be used with the class ItemToBlock, a plain lookup class.